refactor: make CancelableJob reusable
