+++ packages/espressocash_app/lib/features/outgoing_split_key_payments/src/bl/oskp_service.dart
@@ -45,9 +45,11 @@ class OSKPService {
   }) async {
     final status = await _createTx(amount: payment.amount, account: account);
 
-    await _repository.save(payment.copyWith(status: status));
+    final newPayment = payment.copyWith(status: status);
 
-    return payment;
+    await _repository.save(newPayment);
+
+    return newPayment;
   }
 
   Future<OutgoingSplitKeyPayment> cancel(
