+++ packages/espressocash_app/lib/features/outgoing_split_key_payments/src/bl/tx_confirmed_watcher.dart
@@ -6,11 +6,11 @@ import 'package:injectable/injectable.dart';
 import 'package:solana/base58.dart';
 
 import '../../../../core/amount.dart';
+import '../../../../core/cancelable_job.dart';
 import '../../../../core/link_shortener.dart';
 import '../../../../core/single_key_payments.dart';
 import '../../../../core/split_key_payments.dart';
 import '../../models/outgoing_split_key_payment.dart';
-import 'cancelable_job.dart';
 import 'payment_watcher.dart';
 import 'repository.dart';
